requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
cassandra-driver==3.28.0
cassio==0.1.3
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.1
exceptiongroup==1.1.3
frozenlist==1.4.0
geomet==0.2.1.post1
greenlet==3.0.0
idna==3.4
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.306
langsmith==0.0.41
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.6
numpy==1.24.4
openai==0.28.1
packaging==23.2
pydantic==2.4.2
pydantic-core==2.10.1
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.21
tenacity==8.2.3
tiktoken 
tqdm==4.66.1
typing-extensions==4.8.0
typing-inspect==0.9.0
urllib3==2.0.5
yarl==1.9.2
